requirements.txt
flask
tensorflow
pillow
numpy
boto3